third_party/blink/web_tests/platform/generic/fast/harness/error-in-async-test-expected.txt

Test that the window.onerror handler in resources/js-test.js stops asynchronous tests

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Will throw an error soon
Simulating an unexpected error
FAIL Unexpected error: Uncaught Error: An unexpected error
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
